+++ app/build.gradle
@@ -2,7 +2,7 @@ apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
 
 android {
-    compileSdkVersion 30
+    compileSdkVersion 31
     defaultConfig {
         applicationId "com.google.wear.whereami"
         minSdkVersion 26
@@ -29,6 +29,7 @@ dependencies {
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     implementation 'androidx.recyclerview:recyclerview:1.2.1'
     implementation 'androidx.wear:wear:1.1.0'
+    implementation 'androidx.lifecycle:lifecycle-service:2.3.1'
     implementation 'com.google.android.gms:play-services-maps:17.0.1'
     implementation 'androidx.wear:wear-complications-data:1.0.0-alpha19'
     implementation 'androidx.wear:wear-complications-data-source:1.0.0-alpha19'
