+++ app/src/debug/res/values/strings.xml
@@ -1,0 +1,3 @@
+<resources xmlns:xliff="urn:oasis:names:tc:xliff:document:1.2">
+    <string name="tile_preview">Where Am I Tile Preview</string>
+</resources>
